requirements.txt
Pillow>=9.0.0
pypinyin>=0.40.0
requests>=2.25.1
pypandoc>=1.8.0
cairosvg>=2.5.2